Nuke `NamespaceDefinition` object
